extract module: get_object::fixture
