+++ apps/wsb-pilot/tests/s3api/get_object/mod.rs
@@ -1,3 +1,4 @@
+mod fixture;
 mod output;
 mod status;
